A workaround for CH340G driver on OS X
It seems to not wait for USB_CONTROL calls to finish before returning.
As a result, they may interleave with subsequent data transfers, which
can wreak all sorts of havoc, changing baud rate and resetting FIFOs
and whatnot.

We add a short sleep after every ioctl() call. It seems to be sufficient
and is short enough to not matter much and be applied indiscriminately
to all platforms.